6_task Create 2 fragment files
