+++ app/src/main/java/com/example/moviesearch/DetailsFragment.kt
@@ -2,12 +2,12 @@ package com.example.moviesearch
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import kotlinx.android.synthetic.main.activity_details.*
+import kotlinx.android.synthetic.main.fragment_details.*
 
-class DetailsActivity : AppCompatActivity() {
+class DetailsFragment : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_details)
+        setContentView(R.layout.`fragment_details.xml`)
 
         //Получаем наш фильм из переданного бандла
         val film = intent.extras?.get("film") as Film
