+++ app/src/main/res/values/strings.xml
@@ -7,4 +7,6 @@
     <string name="menu_favorites_title">FAVOURITE</string>
     <string name="btn_sett">SETTINGS</string>
     <string name="page_title">Search it!</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
